requirements.txt
fastapi==0.95.0
uvicorn==0.20.0
pydantic==1.10.7
requests==2.28.2
tensorflow==2.12.0  # For DNN implementation
numpy==1.24.2
alembic==1.10.2     # For database migrations
python-dotenv==1.0.0